requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
environs==11.0.0
frozenlist==1.4.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
greenlet==3.1.0
gspread==6.1.2
idna==3.8
magic-filter==1.0.12
marshmallow==3.22.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.11.1
